requirements.txt
Django==5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Pillow==10.0.0
celery==5.3.4
redis==5.0.1
mysqlclient==2.2.0